Combine API routes into one Flask app
Combine API routes into one Flask app, initial groundwork for actually getting the services running. Currently not functional as a whole.
