+++ .gitignore
@@ -8,6 +8,7 @@ Untitled*.ipynb
 *.cache*
 *.conf*
 *temp/*
+*nohup.out
 
 */build/*
 */dist/*
